refactor: BottleLogFile parsing and removed legacy parsing files
