+++ src/ctdam/parser/seabird_data_files.py
@@ -4,6 +4,8 @@ from datetime import datetime, timedelta
 from pathlib import Path
 from typing import Tuple
 
+
+import pandas as pd
 import numpy as np
 import odf.sbe.accessors
 import xarray as xr
@@ -528,3 +530,88 @@ class HexFile(SeabirdDataFile):
             return HexFile(previous_hexes[-1]).xmlcon
 
         return XMLCONFile(xmlcons[0])
+
+
+class BottleLogFile(SeabirdDataFile):
+    """
+    Bottle Log file (.bl) representation, that extracts the three different data
+    types from the file: reset time and the table with bottle IDs and
+    corresponding data ranges.
+
+    Parameters
+    ----------
+    path_to_file : Path | str
+        Path to .bl file
+    create_dataframe : bool
+        Whether to parse as dataframe
+    """
+
+    def __init__(self, path_to_file):
+        super().__init__(path_to_file)
+        self.reset_time = self.obtaining_reset_time()
+        self.data = self.data_whitespace_removal()
+        self.df = self.create_dataframe()
+
+    def data_whitespace_removal(self) -> list:
+        """Strips the input from whitespace characters, in this case especially
+        newline characters.
+
+        Parameters
+        ----------
+
+        Returns
+        -------
+        the original data stripped off the whitespaces
+
+        """
+        temp_data = []
+        for line in self.raw_data[2:]:
+            temp_data.append(line.strip())
+        return temp_data
+
+    def obtaining_reset_time(self) -> datetime:
+        """Reading reset time with small input check.
+
+        Parameters
+        ----------
+
+        Returns
+        -------
+        a datetime.datetime object of the device reset time
+
+        """
+
+        regex_check = re.search(
+            r"RESET\s(\w{3}\s\d+\s\d{4}\s\d\d:\d\d:\d\d)",
+            self.raw_data[1],
+        )
+        if regex_check:
+            return datetime.strptime(regex_check.group(1), "%b %d %Y %H:%M:%S")
+        else:
+            error_message = """BottleLogFile is not formatted as expected:
+                Reset time could not be extracted."""
+            logger.error(error_message)
+            raise IOError(error_message)
+
+    def create_dataframe(self) -> pd.DataFrame:
+        """Creates a dataframe from the list specified in self.data.
+
+        Parameters
+        ----------
+
+        Returns
+        -------
+        a pandas.Dataframe representing the bl files table information
+        """
+        data_lists = []
+        for line in self.data:
+            inner_list = line.split(",")
+            # dropping first column as its the index
+            data_lists.append(inner_list[1:])
+        df = pd.DataFrame(data_lists)
+        df.columns = ["Bottle ID", "Datetime", "start_range", "end_range"]
+        df["Datetime"] = pd.to_datetime(df["Datetime"])
+        df["Bottle ID"].astype("int")
+        df["start_range"].astype("int")
+        df["end_range"].astype("int")
+        return df
